* master: (319 commits)
  [AdminListBundle] [TranslatorBundle] Option to filter on empty fields in StringFilterType. (#2434)
  fix admin test generator (#2432)
  Fix addField docblock param for AbstractAdminListConfigurator (#2433)
  [NodeBundle] Register services that enable the node bundle entity tabs (#2430)
  [GeneratorBundle] Added fos user config to generator (#2427)
  [NodeBundle] Allow entity as creator (#2429)
  [AllBundle] Separate bundle phpunit config and cleanup old files (#2419)
  check if db_driver has been configured as it cannot be overwritten (#2425)
  [AdminBundle] dynamic breakpoint for top navbar in admin area (#2375)
  [AdminBundle] update Bootstrap v3.3 -> v3.4 (#2424)
  [Documentation] KunstmaanCMS Installer documentation (#2418)
  [AdminBundle] fix typo in filename of wordcount (#2423)
  [FormBundle] In newer versions of symfony the @templating service is changed to the delegatingengine instead of the twigengine (#2422)
  [AllBundles] Replace deprecated getRootDir calls (#2421)
  [NodeBundle] Remove old/incorrect doc about lock command (#2420)
  fix null check (#2417)
  [GeneratorBundle] Added config generator + improved install command (#2414)
  [NodeBundle] Import commands.yml to register command in sf4 (#2415)
  [NodeBundle] Fix for missing validation of url/email link chooser (#2413)
  [AllBundles] added link script for development (#2412)
  ...

+++ src/Kunstmaan/TranslatorBundle/Tests/unit/Service/Translator/CacheValidatorTest.php
@@ -1,71 +1,71 @@
 <?php
+
 namespace Kunstmaan\TranslatorBundle\Tests\Service\Importer;
 
-use Kunstmaan\TranslatorBundle\Tests\unit\BaseTestCase;
+use Kunstmaan\TranslatorBundle\Repository\TranslationRepository;
+use Kunstmaan\TranslatorBundle\Service\Translator\CacheValidator;
+use PHPUnit\Framework\TestCase;
 
-class CacheValidatorTest extends BaseTestCase
+class CacheValidatorTest extends TestCase
 {
-
     private $cacheValidator;
-    private $languages;
-    private $cacheDir;
+
+    private $cacheDir = __DIR__.'/../../app/cache';
 
     public function setUp()
     {
-        parent::setUp();
-        $this->cacheValidator = $this->getContainer()->get('kunstmaan_translator.service.translator.cache_validator');
+        $date = new \DateTimeImmutable();
+        $yesterday = $date->modify('-1 day');
 
-        $this->languages = array('nl', 'fr', 'de', 'es');
-        $this->cacheDir = sprintf('%s/translations/', $this->getContainer()->getParameter('kernel.cache_dir'));
+        $translationRepository = $this->createMock(TranslationRepository::class);
+        $translationRepository->method('getLastChangedTranslationDate')->willReturn($yesterday);
+
+        $this->cacheValidator = new CacheValidator();
+        $this->cacheValidator->setTranslationRepository($translationRepository);
+
+        $this->cacheDir = sprintf('%s/translations/', $this->cacheDir);
         if (!is_dir($this->cacheDir)) {
             mkdir($this->cacheDir, 0777, true);
         }
+
+        $this->cacheValidator->setCacheDir($this->cacheDir);
     }
 
-    /**
-     * @group cacher
-     */
-    public function testNoGetOldestCachefileDate()
+    public function tearDown()
     {
-        $data = $this->cacheValidator->getOldestCachefileDate();
-        $this->assertNull($data);
+        $this->deleteDummyCachedFile();
     }
 
     /**
      * @group cacher
      */
-    public function testGetOldestCachefileDate()
-    {
-        $this->createDummyTranslationFiles();
-        $date = $this->cacheValidator->getOldestCachefileDate();
-        $this->deleteDummyTranslationFiles();
-        $this->assertInstanceOf('\DateTime', $date);
-    }
-
-    /**
-     * @group cacher
-     */
-    public function testisCacheFresh()
+    public function testIsCacheFreshWithNoCache()
     {
         // cache is always fresh, because there is none
         $fresh = $this->cacheValidator->isCacheFresh();
         $this->assertTrue($fresh);
     }
 
-    public function createDummyTranslationFiles()
+    /**
+     * @group cacher
+     */
+    public function testIsCacheFreshWithExistingCache()
     {
-        foreach ($this->languages as $language) {
-            $command = sprintf('touch %s/catalogue.%s.php', $this->cacheDir, $language);
-            exec($command);
-        }
+        $this->createDummyCachedFile();
+        $fresh = $this->cacheValidator->isCacheFresh();
+        $this->assertTrue($fresh);
     }
 
-    public function deleteDummyTranslationFiles()
+    public function createDummyCachedFile()
     {
-        foreach ($this->languages as $language) {
-            $file = sprintf('%s/catalogue.%s.php', $this->cacheDir, $language);
+        touch(sprintf('%s/catalogue.test.php', $this->cacheDir));
+    }
+
+    public function deleteDummyCachedFile()
+    {
+        $file = sprintf('%s/catalogue.test.php', $this->cacheDir);
+        if (file_exists($file)) {
             unlink($file);
         }
     }
-
 }
